# contracts/openapi.yaml
...
openapi: "3.1.0"
info:
  title: "材料实验助手 API"
  version: "1.0.0"
  description: |-
    当前后端实现的接口清单；维护依据为 backend 中的 Controller、Service、响应模型和安全配置。
    普通成功响应使用 data 和 request_id；分页另有 meta。空 data 按当前 NON_NULL 配置省略。
    204 和文件正文不使用成功信封。错误返回 application/json，包含业务 Problem Details 兼容对象、精简安全错误或内部错误；不能假定所有错误都有 detail、instance 或 request_id。

    前端使用同源 Session（默认 Cookie 为 JSESSIONID）。先获取 /auth/csrf 并保留 csrftoken，写请求携带匹配的 X-CSRFToken；登录也受 CSRF 校验。当前 SecurityConfig 另启用了 HTTP Basic。
    项目和实验读取按当前组织限定数据范围，普通用户另受项目、负责人或实验参与关系约束。创建实验或变更实验关联项目时先校验项目属于当前组织；超级管理员仅跳过项目管理成员检查。平台管理员仅能开通与查看组织目录，不获得其他组织的业务数据读取权。项目和实验受版本保护的操作使用 ETag / If-Match；当前没有 Idempotency-Key 去重实现。

    时间字段沿用现有 API 的 ISO 8601 表示，带时间的响应为含偏移量的日期时间。规范要求客户端展示及人工记录统一到 Asia/Shanghai，但当前页面部分格式化使用浏览器默认时区，尚未统一。项目和实验计划时间省略偏移量时服务端按 Asia/Shanghai 解释；调用方仍应优先提交含偏移量的日期时间。

    能力边界：ready 检查应用启动就绪及数据库连接；项目文档和实验附件正文保存在 PostgreSQL，文档 preview 返回原件。RustFS 存储和 LibreOffice 服务端转换尚未接入。
    实验参与人可随创建或 PATCH 写入；项目成员仅可读取，创建时加入创建者。开发初始化不能替代生产组织、角色和管理员引导；本契约不表示已完成生产部署验收。
servers:
  -
    url: "/api/v1"
    description: "同源 API"
...
security:
  -
    cookieAuth: []
  -
    basicAuth: []
paths:
...
  "/experiments":
    get:
      operationId: "listExperiments"
      tags: ["Experiments"]
      summary: "分页查询可见实验及电子记录"
      description: "要求 experiment.read 权限；普通用户按所属项目的负责人或成员身份限定可见范围。"
      parameters:
        -
          "$ref": "#/components/parameters/Page"
        -
          "$ref": "#/components/parameters/PageSize"
        -
          "$ref": "#/components/parameters/Search"
        -
          name: "project_id"
          in: "query"
          schema:
            type: "string"
            format: "uuid"
        -
          name: "status"
          in: "query"
          schema:
            type: "string"
          description: "按实验状态原值筛选"
      responses:
        "200":
          description: "含电子记录的实验分页列表"
          content:
            "application/json":
              schema:
                allOf:
                  -
                    "$ref": "#/components/schemas/PageEnvelope"
                  -
                    type: "object"
                    required: ["data"]
                    properties:
                      data:
                        type: "array"
                        items:
                          "$ref": "#/components/schemas/Experiment"
        default:
          "$ref": "#/components/responses/ApiError"
    post:
      operationId: "createExperiment"
      tags: ["Experiments"]
      summary: "创建实验和电子实验记录"
      description: |-
        要求 experiment.create、experiment.read；普通用户须满足目标项目管理范围，负责人须为当前组织有效用户。
        所有用户均须保证目标项目属于当前组织；超级管理员仅跳过项目管理成员检查。
        初始状态固定 not_started，自动生成实验编号，并保存计划时间、参与人和 ELN 字段。
      security:
        -
          cookieAuth: []
          csrfCookie: []
          csrfHeader: []
        -
          basicAuth: []
          csrfCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              allOf:
                -
                  "$ref": "#/components/schemas/ExperimentCreate"
                -
                  "$ref": "#/components/schemas/RecordWrite"
      responses:
        "201":
          "$ref": "#/components/responses/Experiment"
        default:
          "$ref": "#/components/responses/ApiError"
  "/experiments/{experimentKey}":
    parameters:
      -
        "$ref": "#/components/parameters/ExperimentKey"
    get:
      operationId: "getExperiment"
      tags: ["Experiments"]
      summary: "按实验编号获取详情及版本"
      description: "要求 experiment.read 权限及实验可见性；当前不支持以 UUID 替代实验编号。"
      responses:
        "200":
          "$ref": "#/components/responses/Experiment"
        default:
          "$ref": "#/components/responses/ApiError"
    patch:
      operationId: "updateExperiment"
      tags: ["Experiments"]
      summary: "按版本保存实验及电子记录"
      description: |-
        要求 experiment.update、experiment.read 且 can_edit=true；已完成实验返回 403，不允许修订。
        普通用户具备 experiment.create 权限时仍须满足实验写入范围；无此权限时仅负责人或参与人可写，并且不能变更 project_id、owner_id。普通用户变更项目另需新项目管理范围。
        所有用户变更 project_id 时均先校验目标项目属于当前组织；超级管理员仅跳过项目管理成员检查。
        name、project_id、owner_id、experiment_type、phase、purpose 缺省时沿用现值。
        estimated_start/estimated_end 缺省时沿用现值，空串或 null 清空；未带偏移量时按 Asia/Shanghai 解释。
        ELN 和 participant_ids 均按字段增量更新：省略或 null 保留，数组传 []、文本传空串可显式清空。ELN 字段在请求顶层，不是嵌套 record。
      security:
        -
          cookieAuth: []
          csrfCookie: []
          csrfHeader: []
        -
          basicAuth: []
          csrfCookie: []
          csrfHeader: []
      parameters:
        -
          "$ref": "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              allOf:
                -
                  "$ref": "#/components/schemas/ExperimentPatch"
                -
                  "$ref": "#/components/schemas/RecordWrite"
      responses:
        "200":
          "$ref": "#/components/responses/Experiment"
        "412":
          "$ref": "#/components/responses/ApiError"
        "428":
          "$ref": "#/components/responses/ApiError"
        default:
          "$ref": "#/components/responses/ApiError"
  "/experiments/{experimentKey}/copy":
    parameters:
      -
        "$ref": "#/components/parameters/ExperimentKey"
    post:
      operationId: "copyExperiment"
      tags: ["Experiments"]
      summary: "复制实验计划"
      description: "要求 experiment.create、experiment.read 及源项目创建范围；复制基础信息、计划时间、参与人、配方和过程，状态重置为 not_started，结果正文及附件不复制。"
      responses:
        "201":
          "$ref": "#/components/responses/Experiment"
        default:
          "$ref": "#/components/responses/ApiError"
  "/experiments/{experimentKey}/attachments":
    parameters:
      -
        "$ref": "#/components/parameters/ExperimentKey"
    post:
      operationId: "uploadExperimentAttachment"
      tags: ["Experiments"]
      summary: "上传实验附件"
      description: "要求实验可编辑。服务端忽略 multipart MIME，按扩展名白名单和基础文件头校验正文；process_image 仅接受不超过 10 MiB 的 JPG/PNG，result_file 使用项目文档相同的格式白名单且不超过 25 MiB。正文保存到 PostgreSQL，成功后实验版本递增。"
      requestBody:
        required: true
        content:
          "multipart/form-data":
            schema:
              type: "object"
              required: ["file","kind"]
              properties:
                file:
                  type: "string"
                  format: "binary"
                kind:
                  type: "string"
                  enum: ["process_image","result_file"]
      responses:
        "201":
          "$ref": "#/components/responses/Experiment"
        default:
          "$ref": "#/components/responses/ApiError"
  "/experiments/{experimentKey}/attachments/{attachmentId}":
    parameters:
      -
        "$ref": "#/components/parameters/ExperimentKey"
      -
        name: "attachmentId"
        in: "path"
        required: true
        schema:
          type: "string"
          format: "uuid"
    delete:
      operationId: "deleteExperimentAttachment"
      tags: ["Experiments"]
      summary: "删除实验附件"
      description: "要求实验可编辑；只删除指定实验下的附件及其数据库正文，成功后实验版本递增。"
      responses:
        "200":
          "$ref": "#/components/responses/Experiment"
        default:
          "$ref": "#/components/responses/ApiError"
  "/experiments/{experimentKey}/attachments/{attachmentId}/content":
    parameters:
      -
        "$ref": "#/components/parameters/ExperimentKey"
      -
        name: "attachmentId"
        in: "path"
        required: true
        schema:
          type: "string"
          format: "uuid"
    get:
      operationId: "getExperimentAttachmentContent"
      tags: ["Experiments"]
      summary: "读取实验附件正文"
      description: "要求 experiment.read 及实验可见性；响应为原始二进制正文。"
      responses:
        "200":
          description: "附件原始正文"
          content:
            "application/octet-stream":
              schema:
                type: "string"
                format: "binary"
        default:
          "$ref": "#/components/responses/ApiError"
  "/experiments/{experimentKey}/transition":
    parameters:
      -
        "$ref": "#/components/parameters/ExperimentKey"
    post:
      operationId: "transitionExperiment"
      tags: ["Experiments"]
      summary: "按顺序推进实验状态"
      description: |-
        要求 experiment.transition、experiment.read、experiment.update 及实验写入范围。
        仅支持 not_started → in_progress → completed，禁止跳级或回退。
        非法迁移返回 409，版本不匹配返回 412；已完成实验先被只读检查拒绝，返回 403。
        状态迁移设置 started_at 或 completed_at，并递增版本。
      security:
        -
          cookieAuth: []
          csrfCookie: []
          csrfHeader: []
        -
          basicAuth: []
          csrfCookie: []
          csrfHeader: []
      parameters:
        -
          "$ref": "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              type: "object"
              required: ["target_status"]
              properties:
                target_status:
                  type: "string"
                  enum: ["in_progress","completed"]
      responses:
        "200":
          "$ref": "#/components/responses/Experiment"
        "409":
          "$ref": "#/components/responses/ApiError"
        "412":
          "$ref": "#/components/responses/ApiError"
        "428":
          "$ref": "#/components/responses/ApiError"
        default:
          "$ref": "#/components/responses/ApiError"
components:
  securitySchemes:
    cookieAuth:
      type: "apiKey"
      in: "cookie"
      name: "JSESSIONID"
      description: "Servlet Session 默认 Cookie 名；服务器配置可覆盖"
    basicAuth:
      type: "http"
      scheme: "basic"
      description: "当前 SecurityConfig 启用的 HTTP Basic；写请求仍受 CSRF 约束"
    csrfCookie:
      type: "apiKey"
      in: "cookie"
      name: "csrftoken"
      description: "先调用 GET /auth/csrf 获取并保留"
    csrfHeader:
      type: "apiKey"
      in: "header"
      name: "X-CSRFToken"
      description: "与当前 csrftoken Cookie 匹配的令牌"
  parameters:
    Page:
      name: "page"
      in: "query"
      schema:
        type: "integer"
        default: 1
      description: "页码从 1 开始，低于 1 按 1 处理"
    PageSize:
      name: "page_size"
      in: "query"
      schema:
        type: "integer"
        default: 20
      description: "每页数量，服务端归一化到 1 至 100"
    Search:
      name: "search"
      in: "query"
      schema:
        type: "string"
      description: "关键字模糊筛选，具体搜索列以接口查询实现为准"
...
    ExperimentKey:
      name: "experimentKey"
      in: "path"
      required: true
      schema:
        type: "string"
      description: "EXP 实验业务编号，不是 UUID"
    IfMatch:
      name: "If-Match"
      in: "header"
      required: true
      schema:
        type: "string"
        example: "\"1\""
      description: "使用最新 ETag 或带引号的 version 整数。整数解析失败返回 428，版本冲突返回 412；当前缺失请求头由通用异常处理兜底，不能承诺与格式错误同样返回 428。"
  responses:
    ApiError:
      description: "业务或参数错误为 Problem Details 兼容对象；安全过滤器返回 status/code/detail；未分类异常返回内部错误。通常 401 为未认证、403 为权限或 CSRF 失败、404 为资源不可见或不存在。当前媒体类型为 application/json，不能假定所有非法输入均为 400。"
      content:
        "application/json":
          schema:
            anyOf:
              -
                "$ref": "#/components/schemas/ProblemDetail"
              -
                "$ref": "#/components/schemas/SecurityError"
              -
                "$ref": "#/components/schemas/InternalError"
...
    Experiment:
      description: "实验详情、电子记录、编辑权限和最新 ETag"
      headers:
        ETag:
          description: "带双引号的版本整数"
          schema:
            type: "string"
            example: "\"1\""
      content:
        "application/json":
          schema:
            allOf:
              -
                "$ref": "#/components/schemas/Envelope"
              -
                type: "object"
                required: ["data"]
                properties:
                  data:
                    "$ref": "#/components/schemas/Experiment"
...
  schemas:
    Envelope:
      type: "object"
      required: ["request_id"]
      properties:
        request_id:
          type: "string"
          description: "当前响应生成的标识"
          format: "uuid"
      description: "非空业务数据通过 data 返回；空 data 按当前 NON_NULL 配置省略。"
    PageEnvelope:
      allOf:
        -
          "$ref": "#/components/schemas/Envelope"
        -
          type: "object"
          required: ["meta"]
          properties:
            meta:
              "$ref": "#/components/schemas/PageMeta"
    PageMeta:
      type: "object"
      required: ["page","page_size","total","total_pages"]
      properties:
        page:
          type: "integer"
          minimum: 1
        page_size:
          type: "integer"
          minimum: 1
          maximum: 100
        total:
          type: "integer"
          format: "int64"
          minimum: 0
        total_pages:
          type: "integer"
          format: "int64"
          minimum: 1
          description: "即使 total 为 0 也返回 1"
    ProblemDetail:
      type: "object"
      required: ["type","title","status","code","detail","instance","request_id"]
      properties:
        type:
          type: "string"
          const: "about:blank"
        title:
          type: "string"
          description: "HTTP 状态英文名称"
        status:
          type: "integer"
        code:
          type: "string"
          description: "业务错误代码，例如 version_conflict"
        detail:
          type: "string"
          description: "业务错误或校验说明"
        instance:
          type: "string"
          description: "请求 URI"
        request_id:
          type: "string"
          format: "uuid"
    SecurityError:
      type: "object"
      required: ["status","code","detail"]
      properties:
        status:
          type: "integer"
          enum: [401,403]
        code:
          type: "string"
          enum: ["authentication_required","permission_denied"]
        detail:
          type: "string"
          const: "认证或权限校验失败"
      description: "安全过滤器直接返回，不包含 request_id、type 或 instance。"
    InternalError:
      type: "object"
      required: ["type","title","status","code","request_id"]
      properties:
        type:
          type: "string"
          const: "about:blank"
        title:
          type: "string"
          const: "服务器内部错误"
        status:
          type: "integer"
          const: 500
        code:
          type: "string"
          const: "internal_error"
        request_id:
          type: "string"
          format: "uuid"
      description: "未分类异常的兜底响应，不包含 detail 或 instance。"
...
    ExperimentCreate:
      type: "object"
      required: ["project_id","name"]
      properties:
        project_id:
          type: "string"
          description: "所属项目主键"
          format: "uuid"
        name:
          type: "string"
          minLength: 1
        experiment_type:
          type: "string"
          description: "缺省为 research"
        phase:
          type: "string"
          description: "缺省为方案设计"
        purpose:
          type: "string"
          description: "缺省为空串"
        owner_id:
          type: "string"
          description: "缺省为当前用户"
          format: "uuid"
        estimated_start:
          type: ["string","null"]
          description: "ISO 8601 日期时间；未带偏移量时按 Asia/Shanghai 解释"
        estimated_end:
          type: ["string","null"]
          description: "格式同 estimated_start，不能早于 estimated_start"
        participant_ids:
          type: "array"
          items:
            type: "string"
            format: "uuid"
    ExperimentPatch:
      type: "object"
      required: []
      properties:
        project_id:
          type: "string"
          format: "uuid"
        name:
          type: "string"
        experiment_type:
          type: "string"
        phase:
          type: "string"
        purpose:
          type: "string"
        owner_id:
          type: "string"
          format: "uuid"
        participant_ids:
          type: ["array","null"]
          items:
            type: "string"
            format: "uuid"
          description: "省略或 null 保留现值；[] 清空；用户须为当前组织有效用户"
        estimated_start:
          type: ["string","null"]
          description: "ISO 8601 日期时间；未带偏移量时按 Asia/Shanghai 解释；省略保留，空串或 null 清空"
        estimated_end:
          type: ["string","null"]
          description: "与 estimated_start 相同格式及清空语义，不能早于 estimated_start"
    RecordWrite:
      type: "object"
      required: []
      properties:
        formula_columns:
          type: ["array","null"]
          items: {}
          description: "配方列配置；创建时省略或 null 写入 []，PATCH 时省略或 null 保留"
        formula_rows:
          type: ["array","null"]
          items: {}
          description: "配方行数据；创建时省略或 null 写入 []，PATCH 时省略或 null 保留"
        extra_tables:
          type: ["array","null"]
          items: {}
          description: "额外表格；创建时省略或 null 写入 []，PATCH 时省略或 null 保留"
        process_text:
          type: ["string","null"]
          description: "实验过程正文；创建时省略或 null 写入空串，PATCH 时省略或 null 保留"
        extra_processes:
          type: ["array","null"]
          items: {}
          description: "额外过程；创建时省略或 null 写入 []，PATCH 时省略或 null 保留"
        result_text:
          type: ["string","null"]
          description: "实验结果正文；创建时省略或 null 写入空串，PATCH 时省略或 null 保留"
      description: "请求顶层 ELN 字段；每个已提交字段整体替换，未提交字段保留。JSON 字段当前未校验元素结构，本契约按前端正常请求数组形态描述。"
    ExperimentRecord:
      type: "object"
      required: ["formula_columns","formula_rows","extra_tables","process_text","extra_processes","process_images","result_text","result_files"]
      properties:
        formula_columns:
          description: "已存储配方列 JSON；正常页面数据为数组"
        formula_rows:
          description: "已存储配方行 JSON；正常页面数据为数组"
        extra_tables:
          description: "已存储额外表格 JSON；正常页面数据为数组"
        process_text:
          type: "string"
        extra_processes:
          description: "已存储额外过程 JSON；正常页面数据为数组"
        process_images:
          type: "array"
          items:
            "$ref": "#/components/schemas/ExperimentAttachment"
          readOnly: true
        result_text:
          type: "string"
        result_files:
          type: "array"
          items:
            "$ref": "#/components/schemas/ExperimentAttachment"
          readOnly: true
    ExperimentAttachment:
      type: "object"
      required: ["id","name","url","size"]
      properties:
        id:
          type: "string"
          format: "uuid"
        name:
          type: "string"
        url:
          type: "string"
          description: "已授权的后端附件正文 URL"
        size:
          type: "string"
          description: "展示文本，当前以 B 或 KB 表示"
      description: "附件元数据投影；正文保存在 PostgreSQL 并通过授权接口读取。"
    Experiment:
      type: "object"
      required: ["id","experiment_no","name","project_id","project_no","project_name","experiment_type","phase","status","purpose","owner_id","owner_display_name","participant_ids","participant_names","record","can_edit","version","created_at","updated_at"]
      properties:
        id:
          type: "string"
          format: "uuid"
        experiment_no:
          type: "string"
        name:
          type: "string"
        project_id:
          type: "string"
          format: "uuid"
        project_no:
          type: "string"
        project_name:
          type: "string"
        experiment_type:
          type: "string"
        phase:
          type: "string"
        status:
          type: "string"
          enum: ["not_started","in_progress","completed"]
        purpose:
          type: "string"
        estimated_start:
          type: "string"
          format: "date-time"
        estimated_end:
          type: "string"
          format: "date-time"
        started_at:
          type: "string"
          format: "date-time"
        completed_at:
          type: "string"
          format: "date-time"
        owner_id:
          type: "string"
          format: "uuid"
        owner_display_name:
          type: "string"
        participant_ids:
          type: "array"
          items:
            type: "string"
            format: "uuid"
          readOnly: true
        participant_names:
          type: "array"
          items:
            type: "string"
          readOnly: true
        record:
          "$ref": "#/components/schemas/ExperimentRecord"
        can_edit:
          type: "boolean"
          description: "服务端按状态、权限和用户范围计算；completed 始终为 false"
          readOnly: true
        version:
          type: "integer"
          minimum: 1
        created_at:
          type: "string"
          format: "date-time"
        updated_at:
          type: "string"
          format: "date-time"
      description: "未赋值的可空日期时间字段按当前配置省略。"
...